backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
asyncpg==0.29.0
httpx==0.26.0
openai==1.10.0
sentence-transformers==2.3.1
aiosqlite==0.19.0
google-generativeai==0.3.2
